refactor moon.js => Luna class
